refactor(walletd): extract wallet state into new file
